"refactor: split skeletonize pipeline from reusable core helpers"
